+++ AddOns/Prat-3.0/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Prat 3.0
 
-## [3.8.14](https://github.com/sylvanaar/prat-3-0/tree/3.8.14) (2020-06-08)
-[Full Changelog](https://github.com/sylvanaar/prat-3-0/compare/3.8.13...3.8.14)
+## [3.8.19](https://github.com/sylvanaar/prat-3-0/tree/3.8.19) (2020-06-10)
+[Full Changelog](https://github.com/sylvanaar/prat-3-0/compare/3.8.18...3.8.19)
 
-- Memory: Retry the profile load until the channels can be joined  
+- Memory: Handle channel number of 0 when adding channels  
+- Memory: If we are restoring an invalid frame - just close the frame  
+- Memory: Ensure frames are properly faded in after restore, and select the first frame in the dock